+++ packages/react-ui/src/theme/sections/components/navigation.js
@@ -10,6 +10,12 @@ export const navigationOverrides = {
         height: theme.spacing(4),
         transition: 'background-color 300ms cubic-bezier(0.4, 0, 0.2, 1) 0ms',
 
+        [theme.breakpoints.up('sm')]: {
+          // Overrides an unwanted Mui default style
+          '&.MuiButtonBase-root': {
+            minHeight: theme.spacing(4)
+          }
+        },
         '&:focus-visible': {
           // Solves a known Mui issue: https://github.com/mui/material-ui/issues/23747
           backgroundColor: 'transparent',
@@ -34,6 +40,7 @@ export const navigationOverrides = {
         },
         '&.Mui-disabled:empty': {
           height: 0,
+          minHeight: 0,
           padding: 0
         },
         '& .MuiCheckbox-root, & > .MuiSvgIcon-root': {
@@ -42,7 +49,14 @@ export const navigationOverrides = {
       }),
       dense: ({ theme }) => ({
         minHeight: theme.spacing(3),
-        height: theme.spacing(3)
+        height: theme.spacing(3),
+
+        [theme.breakpoints.up('sm')]: {
+          // Overrides an unwanted Mui default style
+          '&.MuiButtonBase-root': {
+            minHeight: theme.spacing(3)
+          }
+        }
       })
     }
   },