+++ packages/react-ui/storybook/stories/atoms/SelectField.stories.js
@@ -1,19 +1,15 @@
 import React, { useState } from 'react';
-import {
-  Box,
-  Chip,
-  FormControl,
-  Grid,
-  InputLabel,
-  MenuItem,
-  OutlinedInput,
-  Select,
-  TextField,
-  styled
-} from '@mui/material';
+import { FormControl, Grid, InputLabel, MenuItem, TextField } from '@mui/material';
 import Typography from '../../../src/components/atoms/Typography';
 import SelectField from '../../../src/components/atoms/SelectField';
-import { Container, Label } from '../../utils/storyStyles';
+import {
+  Container,
+  DocContainer,
+  DocHighlight,
+  DocLink,
+  Label
+} from '../../utils/storyStyles';
+import Button from '../../../src/components/atoms/Button';
 
 const options = {
   title: 'Atoms/Select Field',
@@ -72,16 +68,6 @@ const options = {
 };
 export default options;
 
-const ChipWrapper = styled(Box, {
-  shouldForwardProp: (prop) => prop !== 'small'
-})(({ theme, small }) => ({
-  display: 'flex',
-  flexWrap: 'wrap',
-  gap: theme.spacing(1),
-  overflow: 'auto',
-  height: theme.spacing(small ? 3 : 4)
-}));
-
 const menuItems = [
   {
     label: 'Ten: super large text with overflow',
@@ -97,22 +83,62 @@ const menuItems = [
   }
 ];
 
-const menuItemsLong = [
-  {
-    label: 'table_openstreetmap_pointsofinterest',
-    value: '10Long'
-  },
-  {
-    label: 'Twenty',
-    value: '20'
-  },
-  {
-    label: 'Thirty',
-    value: '30'
-  }
-];
+const SelectFieldItem = ({
+  label,
+  required,
+  placeholder,
+  variant,
+  helperText,
+  size,
+  focused,
+  disabled,
+  error,
+  ...rest
+}) => {
+  const [content, setContent] = useState([]);
 
-const PlaygroundTemplate = (args) => <SelectField {...args} items={menuItems} />;
+  const handleChange = (event) => {
+    const {
+      target: { value }
+    } = event;
+    setContent(
+      // On autofill we get a stringified value
+      typeof value === 'string' ? value.split(',') : value
+    );
+  };
+
+  return (
+    <SelectField
+      {...rest}
+      label={label}
+      variant={variant}
+      placeholder={placeholder}
+      helperText={helperText}
+      onChange={handleChange}
+      value={content}
+      focused={focused}
+      disabled={disabled}
+      error={error}
+      size={size}
+      fullWidth={rest.fullWidth}
+    >
+      {[...Array(20)].map((item, index) => {
+        const itemText =
+          index === 1
+            ? `Very long item text with overflow ${index + 1}`
+            : `Item ${index + 1}`;
+
+        return (
+          <MenuItem key={index} value={itemText}>
+            {itemText}
+          </MenuItem>
+        );
+      })}
+    </SelectField>
+  );
+};
+
+const PlaygroundTemplate = (args) => <SelectFieldItem {...args} />;
 
 const VariantsTemplate = ({ label, required, placeholder, ...rest }) => {
   return (
@@ -120,36 +146,33 @@ const VariantsTemplate = ({ label, required, placeholder, ...rest }) => {
       <Grid item>
         <Container>
           <Label variant='body2'>{'Filled'}</Label>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             label={label}
             variant='filled'
             placeholder={placeholder}
-            items={menuItems}
           />
         </Container>
       </Grid>
       <Grid item>
         <Container>
           <Label variant='body2'>{'Outlined'}</Label>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             label={label}
             variant='outlined'
             placeholder={placeholder}
-            items={menuItems}
           />
         </Container>
       </Grid>
       <Grid item>
         <Container>
           <Label variant='body2'>{'Standard'}</Label>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             label={label}
             variant='standard'
             placeholder={placeholder}
-            items={menuItems}
           />
         </Container>
       </Grid>
@@ -163,41 +186,30 @@ const LabelAndHelperTextTemplate = ({ label, placeholder, helperText, ...rest })
       <Grid item>
         <Container>
           <Label variant='body2'>{'Label + helper text'}</Label>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             label={label}
             placeholder={placeholder}
             helperText={helperText}
-            items={menuItems}
           />
         </Container>
       </Grid>
       <Grid item>
         <Container>
           <Label variant='body2'>{'Without label + helper text'}</Label>
-          <SelectField {...rest} placeholder={placeholder} items={menuItems} />
+          <SelectFieldItem {...rest} placeholder={placeholder} />
         </Container>
       </Grid>
       <Grid item>
         <Container>
           <Label variant='body2'>{'Only label'}</Label>
-          <SelectField
-            {...rest}
-            label={label}
-            placeholder={placeholder}
-            items={menuItems}
-          />
+          <SelectFieldItem {...rest} label={label} placeholder={placeholder} />
         </Container>
       </Grid>
       <Grid item>
         <Container>
           <Label variant='body2'>{'Only helper text'}</Label>
-          <SelectField
-            {...rest}
-            placeholder={placeholder}
-            helperText={helperText}
-            items={menuItems}
-          />
+          <SelectFieldItem {...rest} placeholder={placeholder} helperText={helperText} />
         </Container>
       </Grid>
     </Grid>
@@ -213,7 +225,7 @@ const SizeTemplate = ({
   ...rest
 }) => {
   const [fixedValue, setFixedValue] = useState('Twenty');
-  const [fixedValue2, setFixedValue2] = useState('Ten');
+  const [fixedValue2, setFixedValue2] = useState('Twenty');
   const [fixedValue3, setFixedValue3] = useState('Thirty');
   const handleChange = (event) => {
     setFixedValue(event.target.value);
@@ -233,32 +245,29 @@ const SizeTemplate = ({
           <Typography variant='body2'>Custom component</Typography>
         </Grid>
         <Grid item xs={3}>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             variant='filled'
             label={label}
             placeholder={placeholder}
-            items={menuItems}
             size={size}
           />
         </Grid>
         <Grid item xs={3}>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             variant='outlined'
             label={label}
             placeholder={placeholder}
-            items={menuItems}
             size={size}
           />
         </Grid>
         <Grid item xs={3}>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             variant='standard'
             label={label}
             placeholder={placeholder}
-            items={menuItems}
             size={size}
           />
         </Grid>
@@ -271,37 +280,19 @@ const SizeTemplate = ({
         <Grid item xs={3}>
           <FormControl>
             <InputLabel>{label}</InputLabel>
-            <Select {...rest} variant='filled' size={size}>
-              {menuItems.map((option) => (
-                <MenuItem key={option.label} value={option.label}>
-                  {option.label}
-                </MenuItem>
-              ))}
-            </Select>
+            <SelectFieldItem {...rest} variant='filled' size={size} />
           </FormControl>
         </Grid>
         <Grid item xs={3}>
           <FormControl>
             <InputLabel>{label}</InputLabel>
-            <Select {...rest} variant='outlined' size={size}>
-              {menuItems.map((option) => (
-                <MenuItem key={option.label} value={option.label}>
-                  {option.label}
-                </MenuItem>
-              ))}
-            </Select>
+            <SelectFieldItem {...rest} variant='outlined' size={size} />
           </FormControl>
         </Grid>
         <Grid item xs={3}>
           <FormControl>
             <InputLabel>{label}</InputLabel>
-            <Select {...rest} variant='standard' size={size}>
-              {menuItems.map((option) => (
-                <MenuItem key={option.label} value={option.label}>
-                  {option.label}
-                </MenuItem>
-              ))}
-            </Select>
+            <SelectFieldItem {...rest} variant='standard' size={size} />
           </FormControl>
         </Grid>
       </Grid>
@@ -374,36 +365,33 @@ const SizeTemplate = ({
           <Typography>Focused</Typography>
         </Grid>
         <Grid item xs={3}>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             variant='filled'
             label={label}
             placeholder={placeholder}
             focused
             size={size}
-            items={menuItems}
           />
         </Grid>
         <Grid item xs={3}>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             variant='outlined'
             label={label}
             placeholder={placeholder}
             focused
             size={size}
-            items={menuItems}
           />
         </Grid>
         <Grid item xs={3}>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             variant='standard'
             label={label}
             placeholder={placeholder}
             focused
             size={size}
-            items={menuItems}
           />
         </Grid>
       </Grid>
@@ -413,36 +401,33 @@ const SizeTemplate = ({
           <Typography>Disabled</Typography>
         </Grid>
         <Grid item xs={3}>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             variant='filled'
             label={label}
             placeholder={placeholder}
             disabled
             size={size}
-            items={menuItems}
           />
         </Grid>
         <Grid item xs={3}>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             variant='outlined'
             label={label}
             placeholder={placeholder}
             disabled
             size={size}
-            items={menuItems}
           />
         </Grid>
         <Grid item xs={3}>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             variant='standard'
             label={label}
             placeholder={placeholder}
             disabled
             size={size}
-            items={menuItems}
           />
         </Grid>
       </Grid>
@@ -452,7 +437,7 @@ const SizeTemplate = ({
           <Typography>Error</Typography>
         </Grid>
         <Grid item xs={3}>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             variant='filled'
             label={label}
@@ -460,11 +445,10 @@ const SizeTemplate = ({
             helperText={helperText}
             error
             size={size}
-            items={menuItems}
           />
         </Grid>
         <Grid item xs={3}>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             variant='outlined'
             label={label}
@@ -472,11 +456,10 @@ const SizeTemplate = ({
             helperText={helperText}
             error
             size={size}
-            items={menuItems}
           />
         </Grid>
         <Grid item xs={3}>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             variant='standard'
             label={label}
@@ -484,7 +467,6 @@ const SizeTemplate = ({
             helperText={helperText}
             error
             size={size}
-            items={menuItems}
           />
         </Grid>
       </Grid>
@@ -500,73 +482,26 @@ const MultipleTemplate = ({
   size,
   ...rest
 }) => {
-  const [personName, setPersonName] = React.useState([]);
-  const isSmall = size === 'small';
-
-  const handleChange = (event) => {
-    const {
-      target: { value }
-    } = event;
-    setPersonName(
-      // On autofill we get a stringified value.
-      typeof value === 'string' ? value.split(',') : value
-    );
-  };
-
   return (
-    <Grid container direction='column' spacing={6}>
-      <Grid item>
-        <Container>
-          <Label variant='body2'>{'Default (custom component)'}</Label>
-          <SelectField
-            {...rest}
-            multiple
-            size={size}
-            label={label}
-            placeholder={placeholder}
-            items={menuItems}
-          />
-        </Container>
-      </Grid>
-      <Grid item xs={3}>
-        <Container>
-          <Label variant='body2'>{'Select (with custom chips)'}</Label>
-          <Select
-            {...rest}
-            label={label}
-            size={size}
-            multiple
-            displayEmpty
-            value={personName}
-            onChange={handleChange}
-            input={<OutlinedInput id='select-multiple-chip' label='Chip' />}
-            renderValue={(selected) => {
-              if (selected.length === 0) {
-                return (
-                  <Typography variant={isSmall ? 'body2' : 'body1'} color='text.hint'>
-                    Placeholder
-                  </Typography>
-                );
-              }
-
-              return (
-                <ChipWrapper small={isSmall}>
-                  {selected.map((value) => (
-                    <Chip size={size} color='default' key={value} label={value} />
-                  ))}
-                </ChipWrapper>
-              );
-            }}
-          >
-            {[...Array(10)].map((x, index) => (
-              <MenuItem key={index} value={`Option item ${index + 1}`}>
-                {`Option item ${index + 1}`}
-              </MenuItem>
-            ))}
-          </Select>
-        </Container>
-      </Grid>
-    </Grid>
+    <DocContainer
+      severity='warning'
+      action={
+        <Button
+          color='inherit'
+          variant='outlined'
+          size='small'
+          href='/?path=/docs/atoms-multiple-select-field--playground'
+        >
+          MultipleSelectField story
+        </Button>
+      }
+    >
+      We have a specific
+      <DocLink href='https://github.com/CartoDB/carto-react/blob/master/packages/react-ui/src/components/atoms/MultipleSelectField.js'>
+        MultipleSelectField
+      </DocLink>
+      component to handle this functionality, check the associated story.
+    </DocContainer>
   );
 };
 
@@ -577,12 +512,7 @@ const BehaviorTemplate = ({ label, placeholder, defaultValue, helperText, ...res
         <Label variant='subtitle1'>{'Overflow'}</Label>
         <Container style={{ maxWidth: '440px' }}>
           <Label variant='body2'>{'Default'}</Label>
-          <SelectField
-            {...rest}
-            label={label}
-            placeholder={placeholder}
-            items={menuItemsLong}
-          />
+          <SelectFieldItem {...rest} label={label} placeholder={placeholder} />
         </Container>
       </Grid>
 
@@ -593,20 +523,10 @@ const BehaviorTemplate = ({ label, placeholder, defaultValue, helperText, ...res
 
           <Grid container spacing={2}>
             <Grid item>
-              <SelectField
-                {...rest}
-                label={label}
-                placeholder={placeholder}
-                items={menuItems}
-              />
+              <SelectFieldItem {...rest} label={label} placeholder={placeholder} />
             </Grid>
             <Grid item>
-              <SelectField
-                {...rest}
-                label={label}
-                placeholder={placeholder}
-                items={menuItems}
-              />
+              <SelectFieldItem {...rest} label={label} placeholder={placeholder} />
             </Grid>
           </Grid>
         </Container>
@@ -616,25 +536,41 @@ const BehaviorTemplate = ({ label, placeholder, defaultValue, helperText, ...res
         <Label variant='subtitle1'>{'Width'}</Label>
         <Container>
           <Label variant='body2'>{'Default (fullWidth)'}</Label>
-          <SelectField
-            {...rest}
-            label={label}
-            placeholder={placeholder}
-            items={menuItems}
-          />
+          <SelectFieldItem {...rest} label={label} placeholder={placeholder} />
         </Container>
         <Container>
           <Label variant='body2'>{'No fullWidth'}</Label>
-          <SelectField
+          <SelectFieldItem
             {...rest}
             label={label}
             placeholder={placeholder}
             fullWidth={false}
-            items={menuItems}
           />
         </Container>
       </Grid>
     </Grid>
+  );
+};
+
+const DocTemplate = () => {
+  return (
+    <DocContainer severity='warning'>
+      This component adds the <i>placeholder</i> logic on top of TextField Mui component.
+      <Typography mt={2}>
+        So, instead of <i>{'<TextField select />'}</i> or <i>{'<Select />'}</i>, you
+        should use this one:
+        <DocHighlight component='span'>
+          react-ui/src/components/atoms/SelectField
+        </DocHighlight>
+      </Typography>
+      <Typography mt={2}>
+        For external use:
+        <DocHighlight component='span'>
+          {'import { SelectField } from "@carto/react-ui";'}
+        </DocHighlight>
+        .
+      </Typography>
+    </DocContainer>
   );
 };
 
@@ -643,28 +579,35 @@ const commonArgs = {
   placeholder: 'Placeholder text',
   helperText: 'Helper text.'
 };
+
 const sizeArgs = {
   helperText: 'This is a error message.'
 };
 
-const disabledControlsVariantsArgTypes = {
-  variant: { table: { disable: true } }
-};
-const disabledControlsSizeArgTypes = {
+const disabledControlsArgTypes = {
   variant: { table: { disable: true } },
+  multiple: { table: { disable: true } }
+};
+
+const disabledControlsSizeArgTypes = {
+  ...disabledControlsArgTypes,
   error: { table: { disable: true } },
   defaultValue: { table: { disable: true } }
 };
 
 export const Playground = PlaygroundTemplate.bind({});
 Playground.args = { ...commonArgs };
+Playground.argTypes = disabledControlsArgTypes;
+
+export const Guide = DocTemplate.bind({});
 
 export const Variants = VariantsTemplate.bind({});
 Variants.args = { ...commonArgs };
-Variants.argTypes = disabledControlsVariantsArgTypes;
+Variants.argTypes = disabledControlsArgTypes;
 
 export const LabelAndHelperText = LabelAndHelperTextTemplate.bind({});
 LabelAndHelperText.args = { ...commonArgs };
+LabelAndHelperText.argTypes = disabledControlsArgTypes;
 
 export const Medium = SizeTemplate.bind({});
 Medium.args = { ...commonArgs, ...sizeArgs, size: 'medium' };
@@ -676,6 +619,8 @@ Small.argTypes = disabledControlsSizeArgTypes;
 
 export const MultipleSelection = MultipleTemplate.bind({});
 MultipleSelection.args = { ...commonArgs };
+MultipleSelection.argTypes = disabledControlsArgTypes;
 
 export const Behavior = BehaviorTemplate.bind({});
 Behavior.args = { ...commonArgs };
+Behavior.argTypes = disabledControlsArgTypes;