+++ packages/react-ui/storybook/stories/foundations/Borders.stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Box, Grid } from '@mui/material';
 import { useTheme } from '@mui/material/styles';
 import Typography from '../../../src/components/atoms/Typography';
-import { FilledContainer } from '../../utils/storyStyles';
+import { DocContainer, DocHighlight, FilledContainer } from '../../utils/storyStyles';
 
 const options = {
   title: 'Foundations/Borders',
@@ -60,6 +60,17 @@ const Template = () => {
   );
 };
 
+const DocTemplate = () => {
+  return (
+    <DocContainer severity='warning'>
+      Use is restringed to a few specific values, defined in the design system, which are:
+      <DocHighlight component='span'>0.5, 1, 1.5, 2</DocHighlight>
+    </DocContainer>
+  );
+};
+
 export const Playground = BorderBox.bind({});
+
+export const Guide = DocTemplate.bind({});
 
 export const BorderRadius = Template.bind({});