+++ packages/react-ui/storybook/stories/molecules/Avatar.stories.js
@@ -2,7 +2,14 @@ import React from 'react';
 import { Grid, useTheme } from '@mui/material';
 import { Star } from '@mui/icons-material';
 import { getCartoColorStylePropsForItem, Avatar } from '@carto/react-ui';
-import { BoxContent, TitleContent } from '../../utils/storyStyles';
+import {
+  BoxContent,
+  DocContainer,
+  DocLink,
+  DocHighlight,
+  TitleContent
+} from '../../utils/storyStyles';
+import Typography from '../../../src/components/atoms/Typography';
 
 const options = {
   title: 'Molecules/Avatar',
@@ -304,25 +311,64 @@ const ColorBackgroundTemplate = ({ ...args }) => {
   const theme = useTheme();
 
   return (
-    <Grid container direction='column' spacing={3}>
-      <Grid item>
-        <BoxContent>
-          <TitleContent variant='body2'>{'Carto qualitative bold'}</TitleContent>
-          <Grid container item spacing={6}>
-            {[...Array(15)].map((x, index) => (
-              <Grid item key={index}>
-                <Avatar
-                  {...args}
-                  style={{
-                    ...getCartoColorStylePropsForItem(theme, index)
-                  }}
-                >{`${index + 1}`}</Avatar>
-              </Grid>
-            ))}
-          </Grid>
-        </BoxContent>
+    <>
+      <DocContainer severity='warning'>
+        We have a color function to automatically get the color for each item, it is
+        <DocHighlight component='span'>getCartoColorStylePropsForItem</DocHighlight>. It
+        uses <DocHighlight component='span'>qualitativeBold</DocHighlight> CARTO colors.
+        <Typography mt={2}>Check the code to se an example of how to use it.</Typography>
+      </DocContainer>
+      <Grid container direction='column' spacing={3}>
+        <Grid item>
+          <BoxContent>
+            <TitleContent variant='body2'>{'Carto qualitative bold'}</TitleContent>
+            <Grid container item spacing={6}>
+              {[...Array(15)].map((x, index) => (
+                <Grid item key={index}>
+                  <Avatar
+                    {...args}
+                    style={{
+                      ...getCartoColorStylePropsForItem(theme, index)
+                    }}
+                  >{`${index + 1}`}</Avatar>
+                </Grid>
+              ))}
+            </Grid>
+          </BoxContent>
+        </Grid>
       </Grid>
-    </Grid>
+    </>
+  );
+};
+
+const DocTemplate = () => {
+  return (
+    <DocContainer severity='warning'>
+      We have our own
+      <DocLink href='https://github.com/CartoDB/carto-react/blob/master/packages/react-ui/src/components/molecules/Avatar.js'>
+        Avatar
+      </DocLink>
+      component that uses <i>Mui Avatar</i> and extends it with a new styling prop,
+      <DocHighlight component='span'>size</DocHighlight>.
+      <Typography mt={2}>
+        This component normalize <i>size</i> and allow us to use only the 4 sizes defined
+        by the design system:
+        <DocHighlight component='span'>large, medium, small and xsmall</DocHighlight>
+      </Typography>
+      <Typography mt={2}>
+        Import it from
+        <DocHighlight component='span'>
+          react-ui/src/components/molecules/Avatar
+        </DocHighlight>
+      </Typography>
+      <Typography mt={2}>
+        For external use:
+        <DocHighlight component='span'>
+          {'import { Avatar } from "@carto/react-ui";'}
+        </DocHighlight>
+        .
+      </Typography>
+    </DocContainer>
   );
 };
 
@@ -334,6 +380,8 @@ const disabledSizeArgType = {
 };
 
 export const Playground = Template.bind({});
+
+export const Guide = DocTemplate.bind({});
 
 export const Shape = ShapeTemplate.bind({});
 Shape.argTypes = disabledVariantArgType;